Started decomposing the god-main.cpp file.
